+++ store/v2/root/store.go
@@ -14,7 +14,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/db/memdb"
 	prefixdb "github.com/cosmos/cosmos-sdk/db/prefix"
 	util "github.com/cosmos/cosmos-sdk/internal"
-	"github.com/cosmos/cosmos-sdk/store/cachekv"
 	sdkmaps "github.com/cosmos/cosmos-sdk/store/internal/maps"
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	types "github.com/cosmos/cosmos-sdk/store/v2"
@@ -811,36 +810,6 @@ func (rs *Store) Query(req abci.RequestQuery) (res abci.ResponseQuery) {
 	}
 
 	return res
-}
-
-// GetKVStore implements BasicRootStore.
-func (cs *cacheStore) GetKVStore(key types.StoreKey) types.KVStore {
-	ret, has := cs.substores[key.Name()]
-	if has {
-		return ret
-	}
-	ret = cachekv.NewStore(cs.source.GetKVStore(key))
-	cs.substores[key.Name()] = ret
-	return ret
-}
-
-// Write implements CacheRootStore.
-func (cs *cacheStore) Write() {
-	for skey, sub := range cs.substores {
-		sub.Write()
-		delete(cs.substores, skey)
-	}
-}
-
-// CacheRootStore implements BasicRootStore.
-// This recursively wraps the CacheRootStore in another cache store.
-func (cs *cacheStore) CacheRootStore() types.CacheRootStore {
-	return &cacheStore{
-		source:        cs,
-		substores:     map[string]types.CacheKVStore{},
-		listenerMixin: &listenerMixin{},
-		traceMixin:    &traceMixin{},
-	}
 }
 
 func loadSMT(stateCommitmentTxn dbm.DBReadWriter, root []byte) *smt.Store {
